upgraded to openldap 2.4.41 and added vc14 solution

+++ servers/slapd/config.h
@@ -2,7 +2,7 @@
 /* $OpenLDAP$ */
 /* This work is part of OpenLDAP Software <http://www.openldap.org/>.
  *
- * Copyright 1998-2012 The OpenLDAP Foundation.
+ * Copyright 1998-2015 The OpenLDAP Foundation.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -144,7 +144,7 @@ typedef struct config_args_s {
 		int v_int;
 		unsigned v_uint;
 		long v_long;
-		unsigned long v_ulong;
+		size_t v_ulong;
 		ber_len_t v_ber_t;
 		char *v_string;
 		struct berval v_bv;
